Cleaned data prep, narrowIE parsing, OIE running and clustering, need to clean actual filtering functionality

+++ SORE/my_utils/convert_json_article_to_OIE5.py
@@ -1,4 +1,6 @@
-import json, glob
+import json, glob, os
+from SORE.my_utils import clean_raw_input
+
 
 def write_sentences_to_txt_file(input_dict, output_folder):
     """
@@ -10,13 +12,20 @@ def write_sentences_to_txt_file(input_dict, output_folder):
     processed_files = []
     for doc_id in input_dict:
         output_name = output_folder + doc_id.replace('.', '_') + '.txt'
-        content = input_dict[doc_id]
-        processed_files.append(output_name)
+
         try:
-            with open(output_name, 'w') as f:
-                for sent_id, sentence in content.items():
-                    f.writelines(sentence['sentence'] + '\n')
-            print('Processed ', doc_id, 'to a separate text file for OIE')
+            if os.path.exists(output_name):
+                print("{} already exists, skipping and assuming it's already been processed!".format(output_name))
+            else:
+                content = input_dict[doc_id]
+                content = clean_raw_input.clean_content(content)
+
+                with open(output_name, 'w') as f:
+                    for sent_id, sentence in content.items():
+                        f.writelines(sentence['sentence'] + '\n')
+
+                processed_files.append(output_name)
+                print('Processed ', doc_id, 'to a separate text file for OIE')
         except TypeError:
             print('Something wrong at: ', doc_id)
